+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 ### Latest Release
 
-##### v0.1.0-test
+##### v0.1.1
 
-Initial release meant for demonstration of features and testing.
+- Persisting state and user information after plugin restarts
+- A gmail ID can now be connected to by multiple Mattermost users. However, one Mattermost user can be connected to only one gmail ID at a time
+- New sub-command: `subscriptions` to display labels currently subscribed by the user
+- Minor changes to display statements and fixed behaviour of disconnect command
 
 
 ### Previous Releases
 
-_Currently, there are no previous releases_
+##### v0.1.0-test
+
+- Initial release meant for demonstration of features and testing.